[cmake] automatically fetch submodules
when configuring le_exr check for whether submodules need to be checked
out - if they need to be checked out, do so.

we do this so that only apps that use le_exr have to pay the price of
downloading and checking out these dependencies.

once the submodules have been checked out, other apps may share this
code.